requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.10
Django==5.1.3
django-autoslug==1.9.9
django-crispy-forms==2.3
django-tinymce==4.1.0
djangorestframework==3.15.2
dnspython==2.7.0
pillow==11.0.0
pymongo==4.10.1
sqlparse==0.5.2
tzdata==2024.2
